requirements.txt
pythorhead>=0.19.1
mastodon.py
loguru